requirements.txt
pandas==2.2.2
python-dotenv==1.0.0
